refactor: messageBox component + pokemon folder structure update
