+++ src/web/src/components/Reward.vue
@@ -11,7 +11,7 @@
 <script>
   import { mapGetters } from 'vuex';
   import pokemonMixin from '@/common/mixins/pokemonMixin';
-  import PokeList from './PokemonList.vue';
+  import PokeList from '@/components/pokemon/PokemonList.vue';
 
   export default {
     name: 'Reward',
