+++ src/web/src/components/battle/Battle.vue
@@ -25,46 +25,17 @@
                            :defaultHP="defaultHP"
                            :isHome="true" />
           </div>
-          <div :class="['messageDiv', 'row', (fullscreen)? 'fullscreen': '']">
-              <div id="battleMessage" class="col-md-6 col-xs-6">
-                  {{ message }}
-              </div>
-              <div id="options" class="col-md-6 col-xs-6" style="width: 50%;float:right">
-                  <div v-for="(move,index) in homebattlePokemon.moves"
-                        class="move"
-                        :key="index"
-                        :class="[isAbilityUsedTooMuch(move) ? 'disabledbutton' : '']"
-                        @click.prevent="attack(move.move)"
-                        v-show="isHomePlayerBattlePhase()"
-                        v-if="Object.keys(homebattlePokemon).length && index < 4">
-                         {{ move.move.name }}
-                  </div>
-                  <span style="cursor: pointer; margin: 2%"
-                        @click.prevent="onWalkAway()"
-                        v-show="isHomePlayerBattlePhase()">
-                         walk away <i class="fas fa-walking fa-2x"></i>
-                  </span>
-                  <span style="cursor: pointer; margin: 2%"
-                        @click.prevent="changePokemon()"
-                        v-show="isHomePlayerBattlePhase()">
-                         pokemon <i class="fas fa-exchange-alt fa-2x"></i>
-                  </span>
-                  <div class="row">
-                    <div v-if="getHomePokemon && !Object.keys(homebattlePokemon).length"
-                        v-for="(poke, index) in getHomePokemon":key="index">
-                      <div class="startersDiv">
-                        <Pokemon :info="poke"
-                                 class="col-md-4 col-xs-4"
-                                 :classFlag="true"
-                                 :action-on-click="onPokemonChoosed"
-                                 :disabled="disabled[poke.name]"
-                                 id="starter">
-                       </Pokemon>
-                     </div>
-                    </div>
-                 </div>
-              </div>
-          </div>
+          <MessageBox :message="message"
+                      :fullscreen="fullscreen"
+                      :homebattlePokemon="homebattlePokemon"
+                      :getHomePokemon="getHomePokemon"
+                      :disabled="disabled"
+                      :onPokemonChoosed="onPokemonChoosed"
+                      :changePokemon="changePokemon"
+                      :isHomePlayerBattlePhase="isHomePlayerBattlePhase"
+                      :onWalkAway="onWalkAway"
+                      :isAbilityUsedTooMuch="isAbilityUsedTooMuch"
+                      :attack="attack" />
         </div>
       </div>
     </fullscreen>
@@ -94,8 +65,9 @@
   import ExtraAward from '@/components/modals/ExtraAward';
   import Stats from '@/components/modals/Stats';
   import Confirm from '@/components/modals/Confirm';
-  import Pokemon from '@/components/Pokemon';
+  import Pokemon from '@/components/pokemon/Pokemon';
   import BattlePokemon from '@/components/battle/BattlePokemon';
+  import MessageBox from '@/components/battle/MessageBox';
   import Score from '@/components/battle/Score';
   import fullscreen from 'vue-fullscreen';
   import Vue from 'vue';
@@ -107,7 +79,7 @@
   export default {
      name: 'Battle',
      mixins: [pokemonMixin, battleMixin, uniqueIdGeneratorMixin],
-     components: { Pokemon, PostGame, ExtraAward, Stats, Confirm, Score, BattlePokemon },
+     components: { Pokemon, PostGame, ExtraAward, Stats, Confirm, Score, BattlePokemon, MessageBox },
      data() {
        return {
           fullscreen: false,
