+++ src/web/src/components/pokemon/Pokemon.vue
