+++ src/web/src/components/pokemon/PokemonCard.vue
